Give Otacilio and Raul the last two things they were short of
Two blockers were left after the palette work: SootopolisCity's five-way
scene wanted five exclusive palettes and could afford four, and RAUL had no
graphics id at all.

The first is not a shortage of banks, it is a shortage of *measured* banks.
Four of the twelve reserved slots hold NPC reflections, filled at map load
whether or not anything will ever reflect into them, and a reflection bank is
read only when a sprite drawn from its base palette acquires a reflection.
Measured per owner rather than per map, that scene leaves three of the four
base palettes with nobody on them at all -- the five people standing in it
have colours of their own. So the exclusive-palette allocator may now borrow
a reflection bank it has proved idle, records the loan, and hands it back the
moment an object event arrives that wants the base palette it mirrors,
rebuilding the engine's palette and rehousing the borrower before the
newcomer's sprite exists to reflect. Loans are cleared with the pool at every
map load. Otacilio now holds bank nine there: five unique palettes, no
fallback.

The second is not a shortage of ids either. An object event stores one byte,
but the VAR_OBJ_GFX_ID_x var it reads holds two, and only the assignment back
into the object ever threw the top byte away. Ids from ARAUNA_VIRTUAL_GFX_START
up therefore live in that second byte and resolve through a small Arauna
registry; they are never stored in an object event, so no engine structure
grows, the save layout does not move and NUM_OBJ_EVENT_GFX stays at 239.
RAUL is its first entry, placed with OBJ_EVENT_GFX_VAR_B on his three maps,
each of which names him in the var in its ON_TRANSITION before objects spawn.
That var is also decoration slot twelve, which is safe because every
decoration room rewrites all of them from the save on entry -- an object gfx
var belongs to whoever owns the current map, as vanilla already uses _0 and _F.

He draws his own art beside real Magma grunts instead of borrowing theirs,
and gReflectionEffectPaletteMap's four dead entries now map to themselves so
a character on a pool bank reflects in its own colours rather than the
player's.

+++ src/event_data.c
@@ -188,7 +188,10 @@ bool8 VarSet(u16 id, u16 value)
     return TRUE;
 }
 
-u8 VarGetObjectEventGraphicsId(u8 id)
+// The var is sixteen bits and so is what comes back. Truncating here is what
+// used to make an object gfx var no wider than a graphics id; the callers
+// decide what to do with a value that does not fit in one byte.
+u16 VarGetObjectEventGraphicsId(u8 id)
 {
     return VarGet(VAR_OBJ_GFX_ID_0 + id);
 }
